refactor: 견적서 관련 로직 상위 컴포넌트로 이동
